pyproject.toml
[project]
name = "img2img-turbo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.32.2",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "peft>=0.15.1",
    "pillow>=11.1.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.50.2",
]
